Rename MirraSuite to MirraMunitSuite and add comprehensive documentation
- Rename MirraSuite → MirraMunitSuite for clarity (munit-specific)
- Add full ScalaDoc to MirraMunitSuite and MirraZIOSuite explaining
  SystemUnderTest construction and step-by-step implementation guide
- Mirror scaladoc content into munit.md and zio-test.md docs
