+++ docs/src/main/laika/getting-started.md
@@ -66,5 +66,5 @@ With the algebra and model in place, wire in a real database backend and a test 
 - [Composing repositories](composing-repositories.md) — test multiple repositories together in a single `SystemUnderTest`
 - [Doobie](doobie.md) — `ConnectionIO`-based backend + munit example
 - [Skunk](skunk.md) — `Kleisli[F, Session[F], *]`-based backend
-- [munit + cats-effect](munit.md) — property-test with `MirraSuite`
+- [munit + cats-effect](munit.md) — property-test with `MirraMunitSuite`
 - [ZIO Test](zio-test.md) — property-test with `MirraZIOSuite`
