+++ zio-test/src/main/scala/mirra/MirraZIOSuite.scala
@@ -8,21 +8,65 @@ import zio.test.*
 
 /** ZIO Test integration for Mirra mirror-testing.
   *
-  * Mirrors [[mirra.MirraSuite]] but targets ZIO Test instead of munit/cats-effect.
-  * The effect type is fixed to [[Task]] (`ZIO[Any, Throwable, *]`); bridging to
-  * Cats Effect (e.g. for Doobie) is done in concrete suites via `zio-interop-cats`.
+  * `MirraZIOSuite` is the ZIO Test counterpart of [[mirra.MirraMunitSuite]].  The effect
+  * type is fixed to [[Task]] (`ZIO[Any, Throwable, *]`); bridging to Cats Effect (e.g. for
+  * Doobie) is done in concrete suites via `zio-interop-cats`.
   *
-  * Concrete suites extend this trait as an `object`, supply three abstract types,
-  * implement [[bootstrapSystemUnderTest]], and define `spec` using the helpers:
+  * == What is a SystemUnderTest? ==
+  *
+  * A [[SystemUnderTest]] (SUT) holds two interpreters of your algebra running side-by-side:
+  *
+  *   - The '''real''' interpreter — your actual database-backed implementation
+  *     (e.g. Doobie or Skunk).
+  *   - The '''model''' interpreter — a pure in-memory [[Mirra]] implementation that acts
+  *     as a reference oracle.
+  *
+  * Both interpreters are fused into a single `Alg[PairedEffect]` value using
+  * `SemigroupalK.productK`.  When you call `eval(f)`, the same program `f` is threaded
+  * through both at once, returning `(realResult, modelResult)`.  [[assertMirroring]] then
+  * asserts the two results are equal using ZIO Test's `assertTrue` macro.
+  *
+  * Constructing a [[SystemUnderTest]] requires four things:
+  *   - `initState: MirraState` — the starting in-memory state (e.g. `Universe.zero`).
+  *   - `db: Alg[TransactionEffect]` — the real database interpreter.
+  *   - `model: Alg[MirraEffect]` — the Mirra model interpreter.
+  *   - `tx: TransactionEffect ~> Task` — a natural transformation that runs (and rolls back)
+  *     a single `TransactionEffect` action as a `Task`.  Use `DoobieSupport.rollbackTrans`
+  *     (lifted via `zio-interop-cats`) or `SkunkSupport.rollbackTrans` for built-in helpers.
+  *
+  * == How to implement ==
+  *
+  * Extend this trait as an `object`, follow these steps, and define `spec`:
+  *
+  *   1. '''Supply three abstract types'''
+  *      - `BootstrapContext` — the infrastructure value produced by your ZLayer bootstrap
+  *        (e.g. a running `PostgreSQLContainer`).
+  *      - `MirraState` — the in-memory domain state threaded through the model interpreter
+  *        (e.g. `Universe`).
+  *      - `TransactionEffect[_]` — the real database effect monad
+  *        (e.g. `ConnectionIO` for Doobie, `Kleisli[F, Session[F], *]` for Skunk).
+  *
+  *   2. '''Implement `bootstrapSystemUnderTest`'''
+  *      Return a `ZIO[Scope, Throwable, SystemUnderTest]` that acquires your database
+  *      connection / transactor and constructs a `new SystemUnderTest(initState, db, model, tx)`.
+  *      Use `ZIO.acquireRelease` / `ZLayer.scoped` for resources that need cleanup.
+  *
+  *   3. '''Write property tests using ZIO Test'''
+  *      Define `spec` using `suite(...)` / `test(...)`.  Inside each test, call
+  *      `assertMirroring(context) { x => … }` where `x` is `Alg[PairedEffect]` — your
+  *      algebra running both interpreters simultaneously.  Share expensive resources across
+  *      tests with `.provideShared(layer)`.
   *
   * {{{
   * object MySpec extends MirraZIOSuite[PersonRepository] {
-  *   override type BootstrapContext    = PostgreSQLContainer
-  *   override type MirraState         = Universe
+  *   override type BootstrapContext     = PostgreSQLContainer
+  *   override type MirraState           = Universe
   *   override type TransactionEffect[A] = ConnectionIO[A]
   *
+  *   // Step 2: SystemUnderTest construction
   *   override def bootstrapSystemUnderTest(c: PostgreSQLContainer): ZIO[Scope, Throwable, SystemUnderTest] = …
   *
+  *   // Step 3: property tests
   *   def spec = suite("MySpec")(
   *     test("insert and read") {
   *       ZIO.serviceWithZIO[PostgreSQLContainer] { container =>
@@ -55,7 +99,11 @@ trait MirraZIOSuite[Alg[_[_]]: SemigroupalK] extends ZIOSpecDefault {
   type MirraEffect[A]  = Mirra[MirraState, A]
   type PairedEffect[A] = Tuple2K[TransactionEffect, MirraEffect, A]
 
-  /** Pairs the real database interpreter with the in-memory Mirra model and evaluates programs against both. */
+  /** Pairs the real database interpreter with the in-memory Mirra model and evaluates programs
+    * against both simultaneously.
+    *
+    * See the [[MirraZIOSuite]] scaladoc for a full explanation of the constructor arguments.
+    */
   final class SystemUnderTest(
     initState: MirraState,
     db: Alg[TransactionEffect],
